Move dates.ts into utils/ directory
